Use /dev/tty to access the current terminal
  * In POSIX, the /dev/tty device is always the "controlling
terminal".  As this is the terminal device that we wish to use,
let's just hardcode it instead of pushing it off to the user, who
might screw things up by trying to be clever (I certainly did,
apparently "not" is an inappropriate terminal device, and "a" and
"tty" wound up being ignored arguments).