Use sprintf() to build a single command string
  * Instead of stuffing the console input buffer with individual
pieces of a command string built up over time, use sprintf() to
produce a buffer with the overall command string and push that to
the console in one go.

  * This, technically, introduces a behavior change: If there are
more than 100 frames, files after frame99 will have three (or more,
if things go on long enough) digits instead of two, and will be
ignored by ExtractHDADump, instead of overwriting files starting
at frame00.

  * The fixed-size (80-character) buffer here is amply large: the
base command length is 24 characters, plus up to 10 for the CORB
buffer address, plus up to 5 more for the frame number (unsigned
short is 16 bits, for a maximum of 65535, which is five digits),
plus a NUL for the terminator, gives 40 characters total.  The
other 40 characters are a safety margin in case someone decides to
play sillybuggers.  Or, looking at it the other way, I said "I use
an 80-character terminal.  The pmemsave command is well short of a
full line even with the monitor prompt.  Therefore, 80 characters
of buffer is ample."  And I was right, it is quite literally twice
what was required, and it's stack space on a generous stack, so
there's little point in reducing it.