Extract tty input stuffing from memory dump function
  * Separation of concerns, introduce explanatory function, and so
on.  Basically, dumpMem() is more properly concerned with what the
command is and that it needs to be stuffed to the tty input buffer,
while stuff_tty_input() then is concerned with HOW to stuff the
input buffer.

  * Historically, other commands have been sent to the qemu
monitor via this mechanism.  Having a name and function for it
would prevent code duplication should we need to do so again in
the future.