"Simplify" one if(){if(){if(){}}} chain
  * Right.  "Simplify".  It's checking the first two digits of an
offset (without reference to length) and the first digit of the
data written.

  * Why is this behavior-preserving?  It's matching an offset of
at least two digits, clearly.  If the offset is longer than two
digits (but less than seven digits), then trace_line[tlo+50] can
not be '4' (it will be somewhere within ", 0x").

  * Why is this not behavior-preserving, and why don't we care?
For addresses 0x020xxxx4 and 0x20xxxx4x, without reference to data,
the old rule would fire and the new rule would not.  A PCI region
larger than 32 megabytes is rather larger than is reasonable, so
accesses to one that would trigger the old rule and not the new
are unlikely to be of consequence.