Tighten up CORBLBASE register handling
  * This is the "CORB Lower Base" register, the low 32 bits of the
CORB buffer address.  Pick the register off by its designated
address (rather than a bunch of aliases), and store the value as
a 32-bit integer (instead of as a string).

  * Using an integer instead of a string means that we no longer
have a check for attempting to dump the CORB buffer before the
address is initialized.  This, in theory, should't be a major
problem, as the HD Audio specification defines a default buffer
address of zero, and a driver is unlikely to enable DMA for the
CORB ring buffer before the base address is initialized.

  * As a side-note, this removes the last of the ad-hoc textual
processing of event data.  Everything is now in machine terms
rather than textual terms, so all weird behavior from the textual
processing is now explicitly weird, and can be addressed.