Clean up line-buffer handling
  * The man page for getline(3) on my local system doesn't
describe semantics for a NULL buffer pointer with non-zero length,
but does describe acceptable semantics for NULL buffer pointer and
zero length.  Initialize the buffer length to zero instead of some
arbitrary value.

  * Also, the buffer ends up being NUL-terminated, and we're no
longer poking through it randomly, so we don't need to clear it
between calls to getline().  Remove the calls to memset().