Lock down CORBWP register decoding a bit
  * Offset 0x48 is CORBWP, the Command Output Ring Buffer Write
Pointer.  Offset 0x480 would somewhere in stream descriptor space,
but it would be for stream 32, and the stream number fields are
only four bits wide.  Offsets beyond that are ridiculous.  Lock
the decoding down to just CORBWP.

  * As far as the data test goes, this appears to be looking for a
CORBWP value of 0xff.  This is always in the low eight bits, and
anything higher-order is essentially ignorable (and that's for the
case of a 16-bit or 32-bit write in the first place).  Just check
the low eight bits, and don't worry about any silliness with the
higher-order bits.