Track and honor CORB size
  * The Command Output Ring Buffer has a number of size options,
it isn't necessarily a full 256 entries.  The actual size is set
by a register, and we can (with a couple of caveats) track that.

  * Caveat the first: The register contains both the current size
and a bitmask of valid sizes.  A driver may decide that the
current size is sufficient and thus not write to the register.
This could be compensated for by checking for read events on the
register.

  * Caveat the second: A controller may support only a single CORB
size.  If a driver knows this about a controller, it may decline
to access CORBSIZE at all.  About the only way to deal with this
would be to add a command-line argument to specify the initial
CORB size.