Create a FIFO in /tmp, and delete it on exit.
  * There's an interesting little trick with qemu's pmemsave
command that we can use to good effect, but it requires a named
pipe.  Create one in /tmp, and delete it once we're done.

  * Caveat: If we're killed via SIGINT or similar, atexit() won't
have chance to run, and the FIFO will remain in /tmp.  Until it is
deleted, we'll create a FIFO with a different name.  This
shouldn't be a particularly big deal, given that /tmp is often
deleted automatically, or even set up on a ramdisk in the first
place.  Also, obvious filename scheme is obvious.

  * Oh, and there's no point in having a /closed/ FIFO, so open it
and keep it open.